✨🚀 feat(docker): update Gramps, Owncloud, Nextcloud, Lidarr, Notemark, and Zipline
This commit updates the versions and configurations of various self-hosted applications:

- Gramps: Updated from v24.7.0 to v24.8.0
- Owncloud: Updated from v10.14.0 to v10.15.0, and updated the `tipi_version` to 12
- Nextcloud Mini: Added a new `docker-compose.json` file to include the Nextcloud Mini app and its database
- Lidarr: Updated from v2.4.3 to v2.5.3, and added `created_at` and `updated_at` fields
- Notemark: Updated from v0.13.0 to v0.13.1, and updated the `tipi_version` to 13
- Zipline: Updated from v3.7.9 to v3.7.10, and updated the `tipi_version` to 11

The main reasons for these changes are to keep the applications up-to-date with the latest versions, which often include bug fixes, security updates, and new features. Additionally, the Nextcloud Mini app has been added to provide a more lightweight Nextcloud experience.