🎨 feat(big-bear-omni-tools): add new application configuration and docker-compose (#115)
* 🎨 feat(big-bear-omni-tools): add new application configuration and docker-compose

This commit adds the configuration and Docker Compose setup for the "Omni Tools" application. The key changes include:

- Adds the `config.json` file with details about the application, such as name, port, version, and description.
- Adds the `docker-compose.yml` file to define the service for running the "Omni Tools" application, including Traefik configuration for routing and SSL termination.
- Adds the application logo and description files.

The changes enable the deployment of the "Omni Tools" application as part of the Tipi platform, providing a self-hosted collection of powerful web-based tools for everyday tasks.

* 🖼️ feat: Update big-bear-omni-tools thumbnail

Update the thumbnail image for the big-bear-omni-tools
application. This change provides a more visually
appealing and representative thumbnail for the
application, which will improve the user experience
when browsing the application catalog.

* 🔧 feat(config): update created_at and updated_at timestamps

Updates the `created_at` and `updated_at` timestamps in the `config.json`
file to reflect the current time. This ensures the application's
configuration is up-to-date and accurately reflects the time of the
last update.

* 🎨 feat(big-bear-omni-tools): update app metadata

Update the `config.json` file for the Big Bear Omni Tools app:
- Change the `created_at` and `updated_at` timestamps to more realistic values
- Replace the `logo.jpg` file with a new `logo.png` file

These changes ensure the app metadata is accurate and up-to-date.